requirements.txt
# 核心框架
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# 地理空间数据处理
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.14.0
streamlit-folium>=0.15.0

# 空间分析和统计
scikit-learn>=1.3.0
scipy>=1.11.0
rasterio>=1.3.0

# 数据库连接
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web服务和API
requests>=2.31.0
owslib>=0.29.0

# 数据可视化
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具库
pydeck>=0.8.0
geopy>=2.3.0
pandas
numpy
matplotlib
seaborn
scikit-learn